wellness_backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

# LangChain (stable old version for memory support)
langchain==0.1.20

# Groq integration compatible with LC 0.1.x
langchain-groq==0.1.3

pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1

